+++ incl/KrimzLib/types/vecmath/vec2.h
@@ -3,25 +3,25 @@
 
 namespace kl {
     struct vec2 {
-		double x;
-		double y;
+		float x;
+		float y;
 
 		// Constructors
 		vec2() {
 			x = 0;
 			y = 0;
 		}
-		vec2(double x, double y) {
+		vec2(float x, float y) {
 			this->x = x;
 			this->y = y;
 		}
 		vec2(kl::point p) {
-			x = double(p.x);
-			y = double(p.y);
+			x = float(p.x);
+			y = float(p.y);
 		}
 		vec2(kl::point a, kl::point b) {
-			x = double(b.x - a.x);
-			y = double(b.y - a.y);
+			x = float(b.x - a.x);
+			y = float(b.y - a.y);
 		}
 
 		// Operator overloading
@@ -31,10 +31,10 @@ namespace kl {
 		kl::vec2 operator - (kl::vec2 obj) {
 			return kl::vec2(x - obj.x, y - obj.y);
 		}
-		kl::vec2 operator * (double a) {
+		kl::vec2 operator * (float a) {
 			return kl::vec2(x * a, y * a);
 		}
-		kl::vec2 operator / (double a) {
+		kl::vec2 operator / (float a) {
 			return kl::vec2(x / a, y / a);
 		}
 		bool operator == (kl::vec2 obj) {
@@ -45,18 +45,18 @@ namespace kl {
 		}
 
 		// Returns the vectors length
-		double length() {
+		float length() {
 			return sqrt(x * x + y * y);
 		}
 
 		// Retruns a normalized vector
 		kl::vec2 normalize() {
-			double vecLen = 1 / length();
+			float vecLen = 1 / length();
 			return kl::vec2(x * vecLen, y * vecLen);
 		}
 
 		// Returns the dot product
-		double dot(vec2 a) {
+		float dot(vec2 a) {
 			return x * a.x + y * a.y;
 		}
