+++ src/Testing.cpp
@@ -2,18 +2,18 @@
 
 
 int main() {
-    kl::array<double> a(5);
-    a.runOnEach([](double* elem) {
+    kl::array<float> a(5);
+    a.runOnEach([](float* elem) {
         *elem = kl::random::getInt(0, 10);
     });
 
-    kl::array<double> b(5);
-    b.runOnEach([](double* elem) {
+    kl::array<float> b(5);
+    b.runOnEach([](float* elem) {
         *elem = kl::random::getInt(0, 10);
     });
 
-    kl::array<double> sum(5);
-    sum.runOnEach([&](double* elem, int i) {
+    kl::array<float> sum(5);
+    sum.runOnEach([&](float* elem, int i) {
         *elem = a[i] + b[i];
     });
